Clean up code for final push Cleans up code for final PR
