--- client/src/components/ContactForm.css
@@ -1,23 +1,0 @@
-.contact {
-  height: 100%;
-}
-
-.status {
-  text-align: center;
-  padding-top: 10px;
-}
-.btn {
-  border: solid black 2px;
-  background-color: #2d201c;
-  color: white;
-}
-
-.btn:hover {
-  background: white;
-  color: #2d201c;
-  border: solid black 2px;
-}
-
-#subject {
-  padding-top:7px;
-}
