--- client/src/components/testimonials.css
@@ -1,3 +1,0 @@
-.review_text {
-  padding-top: 10%;
-}
