+++ client/src/index.scss
@@ -3,4 +3,5 @@ body {
   font-family: "Merriweather", serif;
   margin: 0;
   padding: 0;
+  background-color: #fff1de;
 }
